+++ src/main/java/tornadofx/ItemControls.kt
@@ -279,10 +279,7 @@ class LazyTreeItem<T : Any>(
         })
     }
 
-    fun hasChildren(): Boolean {
-        val result = invokeAndSetChildFactorySynchronously()
-        return result == null || result.isEmpty()
-    }
+    fun hasChildren(): Boolean = invokeAndSetChildFactorySynchronously().isNullOrEmpty()
 
     private fun invokeAndSetChildFactorySynchronously(): List<T>? {
         if (!childFactoryInvoked) {
@@ -530,6 +527,12 @@ fun <S> ListView<S>.useCheckbox(converter: StringConverter<S>? = null, getter: (
 }
 
 fun <T> TableView<T>.bindSelected(property: Property<T>) {
+    selectionModel.selectedItemProperty().onChange {
+        property.value = it
+    }
+}
+
+fun <T> ComboBox<T>.bindSelected(property: Property<T>) {
     selectionModel.selectedItemProperty().onChange {
         property.value = it
     }