+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 - `App.scope` is overridable
 - `DefaultScope` deprecated, use `FX.defaultScope` instead
 - The Workspace inside the `scope` of a UIComponents will assume the Workspace it is docked in (https://github.com/edvin/tornadofx/issues/806)
-- Kotlin 1.2.70
+- Kotlin 1.3.0
 - bindSelected for ViewModel gets `out` modifier (https://github.com/edvin/tornadofx/issues/823)
 - Added support for more classes in `EventTarget.properties`.
 - Extracted (some) property keys to top level constants.
@@ -26,6 +26,7 @@
 - Media.play() shortcut which creates a MediaPlayer and plays the Media
 - Wipe and Dissolve view transitions
 - `tab` builder assigns `UIComponent.icon` as Tab graphic
+- `ComboBox.bindSelected()` (https://github.com/edvin/tornadofx/issues/829)
 - Added some function variations with `ObservableValue<T>` instead of `T`. 
     - For an existing `fun build(s: String)` a new variant `fun build(s: ObservableValue<String>)` is added.